+++ src/modules/admin/assets/VideoAsset.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace davidhirtz\yii2\media\video\modules\admin\assets;
+
+use yii\web\AssetBundle;
+
+class VideoAsset extends AssetBundle
+{
+    public $sourcePath = '@vendor/davidhirtz/yii2-media-video/src/modules/admin/assets/video';
+}